requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.2.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.1
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
schedule==1.2.2
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.98
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
zipp==3.21.0
